pass contact identity to custom tools and split AI settings pages
Custom HTTP tools now get the contact's external id and email as
server-side headers, so a tool like a CRM lookup knows who the customer
is. The tabbed AI settings screen is now three pages: providers,
snippets, and tools, each with its own route and sidebar link.
