+++ internal/ai/tools.go
@@ -100,20 +100,29 @@ func (t *searchArticlesTool) Execute(ctx context.Context, args string) (string, 
 	return b.String(), nil
 }
 
+// ToolContext is per-run context injected into custom tool calls server-side. It is never part
+// of the model-facing schema, so the model cannot see or spoof it (e.g. the contact's identity).
+type ToolContext struct {
+	ContactExternalID string
+	ContactEmail      string
+}
+
 // httpTool is an admin-defined custom tool that calls an external HTTP endpoint.
 type httpTool struct {
 	tool          models.Tool
 	encryptionKey string
 	lo            *logf.Logger
 	client        *http.Client
+	tctx          ToolContext
 }
 
-func newHTTPTool(t models.Tool, encryptionKey string, lo *logf.Logger) *httpTool {
+func newHTTPTool(t models.Tool, encryptionKey string, lo *logf.Logger, tctx ToolContext) *httpTool {
 	return &httpTool{
 		tool:          t,
 		encryptionKey: encryptionKey,
 		lo:            lo,
 		client:        toolHTTPClient,
+		tctx:          tctx,
 	}
 }
 
@@ -157,6 +166,15 @@ func (t *httpTool) Execute(ctx context.Context, args string) (string, error) {
 		}
 	}
 
+	// Inject the contact's identity so tools (e.g. a CRM lookup) know who this is. Server-side
+	// only, never advertised to the model.
+	if t.tctx.ContactExternalID != "" {
+		req.Header.Set("X-Libredesk-Contact-External-Id", t.tctx.ContactExternalID)
+	}
+	if t.tctx.ContactEmail != "" {
+		req.Header.Set("X-Libredesk-Contact-Email", t.tctx.ContactEmail)
+	}
+
 	resp, err := t.client.Do(req)
 	if err != nil {
 		t.lo.Error("error calling custom tool", "tool", t.tool.Name, "error", err)
@@ -172,7 +190,7 @@ func (t *httpTool) Execute(ctx context.Context, args string) (string, error) {
 }
 
 // buildToolRegistry returns the built-in, request-injected, and enabled custom tools plus their model-facing definitions.
-func (m *Manager) buildToolRegistry(extra ...Tool) (map[string]Tool, []models.ToolDef, error) {
+func (m *Manager) buildToolRegistry(tctx ToolContext, extra ...Tool) (map[string]Tool, []models.ToolDef, error) {
 	registry := map[string]Tool{}
 	var defs []models.ToolDef
 
@@ -192,7 +210,7 @@ func (m *Manager) buildToolRegistry(extra ...Tool) (map[string]Tool, []models.To
 			m.lo.Warn("skipping custom tool that collides with a built-in tool", "name", ct.Name)
 			continue
 		}
-		ht := newHTTPTool(ct, m.encryptionKey, m.lo)
+		ht := newHTTPTool(ct, m.encryptionKey, m.lo, tctx)
 		registry[ht.Name()] = ht
 		defs = append(defs, toolDef(ht))
 	}
